[level 1] Title: 정수 제곱근 판별, Time: 0.20 ms, Memory: 76.2 MB -BaekjoonHub

+++ 프로그래머스/1/12934. 정수 제곱근 판별/정수 제곱근 판별.java
@@ -3,7 +3,7 @@ class Solution {
         long answer = -1;
         Double sqrt = Math.sqrt(n);
 
-        if(sqrt == sqrt.intValue()) {
+        if(sqrt == sqrt.longValue()) {
             answer = (long) Math.pow(sqrt+1,2);
         }
